+++ chaincode/computeplan/contract.go
@@ -155,7 +155,7 @@ func (s *SmartContract) GetEvaluateTransactions() []string {
 	return commonserv.ReadOnlyMethods["ComputePlan"]
 }
 
-// UpdatePlan updates an Compute Plan in world state
+// UpdatePlan updates a Compute Plan in world state
 // If the key does not exist, it will throw an error
 func (s *SmartContract) UpdatePlan(ctx ledger.TransactionContext, wrapper *communication.Wrapper) error {
 	provider, err := ctx.GetProvider()
@@ -184,4 +184,34 @@ func (s *SmartContract) UpdatePlan(ctx ledger.TransactionContext, wrapper *commu
 	}
 
 	return nil
+}
+
+func (s *SmartContract) IsPlanRunning(ctx ledger.TransactionContext, wrapper *communication.Wrapper) (*communication.Wrapper, error) {
+	provider, err := ctx.GetProvider()
+	if err != nil {
+		return nil, err
+	}
+	service := provider.GetComputePlanService()
+
+	param := new(asset.IsPlanRunningParam)
+	err = wrapper.Unwrap(param)
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to unwrap param")
+		return nil, err
+	}
+
+	isRunning, err := service.IsPlanRunning(param.Key)
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to get compute plan")
+		return nil, err
+	}
+
+	resp := &asset.IsPlanRunningResponse{IsRunning: isRunning}
+
+	wrapped, err := communication.Wrap(ctx.GetContext(), resp)
+	if err != nil {
+		s.logger.Error().Err(err).Msg("failed to wrap response")
+		return nil, err
+	}
+	return wrapped, nil
 }