+++ chaincode/computeplan/contract_test.go
@@ -111,6 +111,7 @@ func TestEvaluateTransactions(t *testing.T) {
 	queries := []string{
 		"GetPlan",
 		"QueryPlans",
+		"IsPlanRunning",
 	}
 
 	assert.Equal(t, queries, contract.GetEvaluateTransactions(), "All non-commit transactions should be flagged")